Move files for the current minicart implementation to /legacy
