+++ react/MinicartLegacy.js
@@ -18,21 +18,21 @@ import { addToCart, updateItems } from 'vtex.store-resources/Mutations'
 import { usePixel } from 'vtex.pixel-manager/PixelContext'
 import ProductPrice from 'vtex.store-components/ProductPrice'
 
-import MiniCartContent from './components/MiniCartContent'
-import Sidebar from './components/Sidebar'
-import Popup from './components/Popup'
-import { shouldShowItem } from './utils/itemsHelper'
-import { mapBuyButtonItemToPixel } from './utils/pixelHelper'
+import MiniCartContent from './legacy/components/MiniCartContent'
+import Sidebar from './legacy/components/Sidebar'
+import Popup from './legacy/components/Popup'
+import { shouldShowItem } from './legacy/utils/itemsHelper'
+import { mapBuyButtonItemToPixel } from './legacy/utils/pixelHelper'
 
-import fullMinicartQuery from './localState/graphql/fullMinicartQuery.gql'
-import updateItemsMutation from './localState/graphql/updateItemsMutation.gql'
-import updateOrderFormMutation from './localState/graphql/updateOrderFormMutation.gql'
-import updateLocalItemStatusMutation from './localState/graphql/updateLocalItemStatusMutation.gql'
-import setMinicartOpenMutation from './localState/graphql/setMinicartOpenMutation.gql'
+import fullMinicartQuery from './legacy/localState/graphql/fullMinicartQuery.gql'
+import updateItemsMutation from './legacy/localState/graphql/updateItemsMutation.gql'
+import updateOrderFormMutation from './legacy/localState/graphql/updateOrderFormMutation.gql'
+import updateLocalItemStatusMutation from './legacy/localState/graphql/updateLocalItemStatusMutation.gql'
+import setMinicartOpenMutation from './legacy/localState/graphql/setMinicartOpenMutation.gql'
 
-import createLocalState, { ITEMS_STATUS } from './localState'
+import createLocalState, { ITEMS_STATUS } from './legacy/localState'
 
-import styles from './minicart.css'
+import styles from './legacy/minicart.css'
 
 const DEFAULT_LABEL_CLASSES = ''
 const DEFAULT_ICON_CLASSES = 'gray'
