+++ react/legacy/__mocks__/react-device-detect.js
