+++ react/legacy/__mocks__/vtex.pixel-manager/PixelContext.js
