+++ react/legacy/__mocks__/vtex.react-portal/Overlay.js
