+++ react/legacy/__mocks__/vtex.react-portal/index.js
