+++ react/legacy/__mocks__/vtex.render-runtime.js
