+++ react/legacy/__mocks__/vtex.store-resources/Mutations.js
