+++ react/legacy/__mocks__/vtex.store-resources/Queries.js
