+++ react/legacy/__tests__/MiniCart.test.js
@@ -3,7 +3,7 @@ import { render, fireEvent } from '@vtex/test-tools/react'
 import { orderForm as orderFormQuery } from 'vtex.store-resources/Queries'
 
 import orderForm from '../__mocks__/orderForm'
-import MiniCart from './../index'
+import MiniCart from '../../MinicartLegacy'
 
 const mocks = [
   {
@@ -163,8 +163,8 @@ describe('<MiniCart />', () => {
   })
 
   it('should not show price if showPrice is false', async () => {
-    const {queryByTestId} = render(
-      <MiniCart type="sidebar" hideContent={false} showPrice={false}/>,
+    const { queryByTestId } = render(
+      <MiniCart type="sidebar" hideContent={false} showPrice={false} />,
       { graphql: { mocks } }
     )
     await flushPromises()
@@ -173,15 +173,15 @@ describe('<MiniCart />', () => {
   })
 
   it('should not show price if there are no items in the cart', () => {
-    const {queryByTestId} = render(
-      <MiniCart type="sidebar" hideContent={false} showPrice={true}/>
+    const { queryByTestId } = render(
+      <MiniCart type="sidebar" hideContent={false} showPrice={true} />
     )
     expect(queryByTestId('total-price')).toBeNull()
   })
 
   it('should show price if there are items in the cart', async () => {
-    const {getByTestId} = render(
-      <MiniCart type="sidebar" hideContent={false} showPrice={true}/>,
+    const { getByTestId } = render(
+      <MiniCart type="sidebar" hideContent={false} showPrice={true} />,
       { graphql: { mocks } }
     )
     await flushPromises()
@@ -190,8 +190,8 @@ describe('<MiniCart />', () => {
   })
 
   it('should not show price if showPrice is false', async () => {
-    const {queryByTestId} = render(
-      <MiniCart type="sidebar" hideContent={false} showPrice={false}/>,
+    const { queryByTestId } = render(
+      <MiniCart type="sidebar" hideContent={false} showPrice={false} />,
       { graphql: { mocks } }
     )
     await flushPromises()
@@ -200,15 +200,15 @@ describe('<MiniCart />', () => {
   })
 
   it('should not show price if there are no items in the cart', () => {
-    const {queryByTestId} = render(
-      <MiniCart type="sidebar" hideContent={false} showPrice={true}/>
+    const { queryByTestId } = render(
+      <MiniCart type="sidebar" hideContent={false} showPrice={true} />
     )
     expect(queryByTestId('total-price')).toBeNull()
   })
 
   it('should show price if there are items in the cart', async () => {
-    const {getByTestId} = render(
-      <MiniCart type="sidebar" hideContent={false} showPrice={true}/>,
+    const { getByTestId } = render(
+      <MiniCart type="sidebar" hideContent={false} showPrice={true} />,
       { graphql: { mocks } }
     )
     await flushPromises()
