+++ react/legacy/components/MiniCartContent.js
