+++ react/legacy/components/MiniCartFooter.js
