+++ react/legacy/components/Sidebar.js
