+++ react/legacy/localState/graphql/fullMinicartQuery.gql
