+++ react/legacy/localState/graphql/minicartIsOpenQuery.gql
