+++ react/legacy/localState/graphql/minicartItemsQuery.gql
