+++ react/legacy/localState/graphql/setMinicartOpenMutation.gql
