+++ react/legacy/localState/graphql/updateLocalItemsMutation.gql
