+++ react/legacy/utils/itemsHelper.js
