+++ react/legacy/utils/pixelHelper.ts
